Refa: split services about llm.
